[Refactor] Club refactoring before API connecting
